Namespace the generic work indexer to CHF
